Rename dot-pe to dot_pe and update all imports to use relative imports
